convert all int to long for time

+++ philo_one/srcs/philosophers.h
@@ -6,7 +6,7 @@
 /*   By: pminne <pminne@student.42lyon.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/09/07 17:15:17 by pminne            #+#    #+#             */
-/*   Updated: 2020/09/19 13:48:34 by pminne           ###   ########lyon.fr   */
+/*   Updated: 2020/10/01 15:49:56 by pminne           ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -37,10 +37,10 @@ int		g_meals_are_done;
 typedef	struct	s_args
 {
 	int		nb_philo;
-	int		time_to_starve;
-	int		time_to_eat;
-	int		time_to_sleep;
-	int		n_time_must_eat;
+	long	time_to_starve;
+	long	time_to_eat;
+	long	time_to_sleep;
+	long	n_time_must_eat;
 }				t_args;
 
 typedef struct	s_write
@@ -68,9 +68,9 @@ typedef struct	s_table
 	pthread_t		th;
 	pthread_t		th_meal;
 	long			start_program;
-	int				time_to_eat;
-	int				time_to_sleep;
-	int				time_to_starve;
+	long			time_to_eat;
+	long			time_to_sleep;
+	long			time_to_starve;
 	long			last_meal;
 	int				nb_philo;
 	int				other_hand;
@@ -87,7 +87,7 @@ t_args			*parsing(int ac, char **av);
 /*
 **		UTILS.C
 */
-int				ft_atoi(char *s);
+long			ft_atoi(char *s);
 unsigned int	ft_strlen(char *s);
 void			ft_putnbr_fd(long n, int fd);
 /*
